# [android] reduce JNI calls by moving logic into Java