--- a/src/Core/AndroidNative/gradle/wrapper/gradle-wrapper.properties
+++ b/src/Core/AndroidNative/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Thu Nov 11 11:45:24 CST 2021
+distributionBase=GRADLE_USER_HOME
+distributionUrl=https\://services.gradle.org/distributions/gradle-7.0.2-bin.zip
+distributionPath=wrapper/dists
+zipStorePath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME