--- a/src/Core/AndroidNative/maui/src/main/java/com/microsoft/maui/ViewHelper.java
+++ b/src/Core/AndroidNative/maui/src/main/java/com/microsoft/maui/ViewHelper.java
@@ -1,0 +1,72 @@
+package com.microsoft.maui;
+
+import android.view.View;
+import android.view.ViewGroup;
+import android.view.ViewParent;
+
+public class ViewHelper {
+    public static void requestLayoutIfNeeded(View view)
+    {
+        if (!view.isInLayout())
+            view.requestLayout();
+    }
+
+    public static void removeFromParent(View view)
+    {
+        ViewParent parent = view.getParent();
+        if (parent == null)
+            return;
+        ((ViewGroup)parent).removeView(view);
+    }
+
+    public static void setPivotXIfNeeded(View view, float pivotX)
+    {
+        if (view.getPivotX() != pivotX)
+            view.setPivotX(pivotX);
+    }
+
+    public static void setPivotYIfNeeded(View view, float pivotY)
+    {
+        if (view.getPivotY() != pivotY)
+            view.setPivotY(pivotY);
+    }
+
+    public static void set(
+        View view,
+        int automationTagId,
+        String automationId,
+        int visibility,
+        int layoutDirection,
+        int minimumHeight,
+        int minimumWidth,
+        boolean enabled,
+        float alpha,
+        float translationX,
+        float translationY,
+        float scaleX,
+        float scaleY,
+        float rotation,
+        float rotationX,
+        float rotationY,
+        float pivotX,
+        float pivotY)
+    {
+        requestLayoutIfNeeded(view);
+        view.setTag(automationTagId, automationId);
+        view.setVisibility(visibility);
+        view.setLayoutDirection(layoutDirection);
+        view.setMinimumHeight(minimumHeight);
+        view.setMinimumWidth(minimumWidth);
+        view.setEnabled(enabled);
+        view.setAlpha(alpha);
+        view.setTranslationX(translationX);
+        view.setTranslationY(translationY);
+        view.setScaleX(scaleX);
+        view.setScaleY(scaleY);
+        view.setRotation(rotation);
+        view.setRotationX(rotationX);
+        view.setRotationY(rotationY);
+        setPivotXIfNeeded(view, pivotX);
+        setPivotYIfNeeded(view, pivotY);
+    }
+}