--- a/src/Core/tests/Benchmarks.Droid/Benchmarks.Droid.csproj
+++ b/src/Core/tests/Benchmarks.Droid/Benchmarks.Droid.csproj
@@ -1,0 +1,53 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFramework>net6.0-android</TargetFramework>
+    <SupportedOSPlatformVersion>21</SupportedOSPlatformVersion>
+    <OutputType>Exe</OutputType>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <ApplicationId>com.microsoft.maui.benchmarks</ApplicationId>
+    <ApplicationVersion>1</ApplicationVersion>
+    <ApplicationDisplayVersion>1.0</ApplicationDisplayVersion>
+    <PublishTrimmed>false</PublishTrimmed>
+    <!-- Physical device is recommended -->
+    <RuntimeIdentifier>android-arm64</RuntimeIdentifier>
+    <!-- HACK: to workaround System.IO.FileNotFoundException: /System.Private.CoreLib.dll -->
+    <EmbedAssembliesIntoApk>false</EmbedAssembliesIntoApk>
+    <AndroidPackageFormat>apk</AndroidPackageFormat>
+  </PropertyGroup>
+  <ItemGroup>
+    <Using Include="Android.Runtime" />
+    <Using Include="Android.Util" />
+    <Using Include="BenchmarkDotNet.Analysers" />
+    <Using Include="BenchmarkDotNet.Attributes" />
+    <Using Include="BenchmarkDotNet.Configs" />
+    <Using Include="BenchmarkDotNet.Exporters" />
+    <Using Include="BenchmarkDotNet.Jobs" />
+    <Using Include="BenchmarkDotNet.Loggers" />
+    <Using Include="BenchmarkDotNet.Order" />
+    <Using Include="BenchmarkDotNet.Running" />
+    <PackageReference Include="BenchmarkDotNet" Version="0.13.1" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\Controls\src\Core\Controls.Core-net6.csproj" />
+    <ProjectReference Include="..\..\src\Core-net6.csproj" />
+  </ItemGroup>
+  <Target Name="Benchmark" DependsOnTargets="Install">
+    <Message Text="Running benchmarks. This might take a while... See 'adb logcat' for realtime progress." Importance="High" />
+    <!-- Clear the log -->
+    <Exec Command="adb shell logcat -c" WorkingDirectory="$(AndroidSdkDirectory)/platform-tools" />
+    <!-- Run benchmarks -->
+    <Exec Command="adb shell am instrument -w $(ApplicationId)/com.microsoft.maui.MainInstrumentation" WorkingDirectory="$(AndroidSdkDirectory)/platform-tools" />
+    <!--
+      Print the log filtered by MAUI and DOTNET tags
+      See: https://developer.android.com/studio/command-line/logcat#outputFormat
+    -->
+    <Exec
+        Command="adb shell logcat -d -v tag -s &quot;DOTNET,MAUI&quot;"
+        IgnoreStandardErrorWarningFormat="true"
+        StdErrEncoding="utf-8"
+        StdOutEncoding="utf-8"
+        WorkingDirectory="$(AndroidSdkDirectory)/platform-tools"
+    />
+  </Target>
+</Project>