… src/Core/AndroidNative/README.md
… src/Core/AndroidNative/gradle/wrapper/gradle-wrapper.jar
… src/Core/AndroidNative/maui/.gitignore
… src/Core/AndroidNative/maui/src/main/AndroidManifest.xml
… src/Core/src/Transforms/Metadata.xml
… src/Core/src/maui.aar
… src/Core/tests/Benchmarks.Droid/AndroidManifest.xml
… src/Core/tests/Benchmarks.Droid/MainActivity.cs
… src/Core/tests/Benchmarks.Droid/MainInstrumentation.cs
… src/Core/tests/Benchmarks.Droid/README.md
… src/Core/tests/Benchmarks.Droid/Resources/layout/activity_main.xml
… src/Core/tests/Benchmarks.Droid/Resources/mipmap-anydpi-v26/ic_launcher.xml
… src/Core/tests/Benchmarks.Droid/Resources/mipmap-anydpi-v26/ic_launcher_round.xml
… src/Core/tests/Benchmarks.Droid/Resources/mipmap-hdpi/ic_launcher.png
… src/Core/tests/Benchmarks.Droid/Resources/mipmap-hdpi/ic_launcher_foreground.png
… src/Core/tests/Benchmarks.Droid/Resources/mipmap-hdpi/ic_launcher_round.png
… src/Core/tests/Benchmarks.Droid/Resources/mipmap-mdpi/ic_launcher.png
… src/Core/tests/Benchmarks.Droid/Resources/mipmap-mdpi/ic_launcher_foreground.png
… src/Core/tests/Benchmarks.Droid/Resources/mipmap-mdpi/ic_launcher_round.png
… src/Core/tests/Benchmarks.Droid/Resources/mipmap-xhdpi/ic_launcher.png
… src/Core/tests/Benchmarks.Droid/Resources/mipmap-xhdpi/ic_launcher_foreground.png
… src/Core/tests/Benchmarks.Droid/Resources/mipmap-xhdpi/ic_launcher_round.png
… src/Core/tests/Benchmarks.Droid/Resources/mipmap-xxhdpi/ic_launcher.png
… src/Core/tests/Benchmarks.Droid/Resources/mipmap-xxhdpi/ic_launcher_foreground.png
… src/Core/tests/Benchmarks.Droid/Resources/mipmap-xxhdpi/ic_launcher_round.png
… src/Core/tests/Benchmarks.Droid/Resources/mipmap-xxxhdpi/ic_launcher.png
… src/Core/tests/Benchmarks.Droid/Resources/mipmap-xxxhdpi/ic_launcher_foreground.png
… src/Core/tests/Benchmarks.Droid/Resources/mipmap-xxxhdpi/ic_launcher_round.png
… src/Core/tests/Benchmarks.Droid/Resources/values/strings.xml
… src/Core/tests/Benchmarks.Droid/ViewHandlerBenchmark.cs