### Description of Change ###

Xamarin.Forms had `FormsViewGroup.java`, which is still used for .NET
MAUI's renderer compatiblity.

The idea, is you could call a single method that sets many values.
This would cross from C# into Java once, instead of several times:

    public void sendBatchUpdate (
            float pivotX,
            float pivotY,
            int visibility,
            boolean enabled,
            float opacity,
            float rotation,
            float rotationX,
            float rotationY,
            float scaleX,
            float scaleY,
            float translationX,
            float translationY){
        setPivotX (pivotX);
        setPivotY (pivotY);

        if (getVisibility () != visibility)
            setVisibility (visibility);

        if (isEnabled () != enabled)
            setEnabled (enabled);

        setAlpha (opacity);
        setRotation (rotation);
        setRotationX (rotationX);
        setRotationY (rotationY);
        setScaleX (scaleX);
        setScaleY (scaleY);
        setTranslationX (translationX);
        setTranslationY (translationY);
    }

One of the drawbacks here was having to subclass `FormsViewGroup`, as
this could have easily been a `static` method for flexibility.

I wrote a benchmark for this idea, and still appears to help
dramatically in .NET 6:

https://github.com/jonathanpeppers/android-pipe

To keep this idea going, let's create `maui.aar` and `ViewHelper.java`:

* `set()` sets many values at once.
* `requestLayoutIfNeeded()` checks `isInLayout()` and calls
  `requestLayout()`.
* `removeFromParent()` does various checks and calls `removeView()`.
* `setPivotXIfNeeded()` (or Y) compares the existing value before setting it.

### Results ###

Running the `Benchmarks.Droid.csproj` on a Pixel 5:

    > ./bin/dotnet/dotnet build src/Core/tests/Benchmarks.Droid/Benchmarks.Droid.csproj -t:Benchmark -c Release

|              Method |       Mean |    Error |   StdDev |  Gen 0 | Allocated |
|-------------------- |-----------:|---------:|---------:|-------:|----------:|
|            --Border |   323.2 µs |  0.82 µs |  0.68 µs | 0.9766 |      5 KB |
|            ++Border |   242.3 µs |  1.34 µs |  1.25 µs | 0.9766 |      5 KB |
|       --ContentView |   354.6 µs |  2.61 µs |  2.31 µs | 1.4648 |      6 KB |
|       ++ContentView |   258.3 µs |  0.49 µs |  0.43 µs | 1.4648 |      6 KB |
| --ActivityIndicator |   486.1 µs |  1.41 µs |  1.18 µs | 0.9766 |      5 KB |
| ++ActivityIndicator |   380.3 µs |  1.40 µs |  1.31 µs | 0.9766 |      5 KB |
|             --Label |   630.8 µs |  3.02 µs |  2.83 µs | 0.9766 |      5 KB |
|             ++Label |   648.8 µs |  6.14 µs |  5.74 µs | 0.9766 |      5 KB |
|             --Entry | 2,169.2 µs | 18.02 µs | 16.85 µs |      - |     12 KB |
|             ++Entry | 1,998.2 µs | 17.01 µs | 15.07 µs |      - |     12 KB |

-- is *before*, with `git checkout head~1 src\Core\src` on this branch.
++ is using the code changes on this branch as-is.

Attaching each of these handlers to native controls is about 30%
improved.

Running the `Maui.Controls.Sample.SingleProject` app:

    .\bin\dotnet\dotnet build .\src\Controls\samples\Controls.Sample.SingleProject\Maui.Controls.Sample.SingleProject.csproj -c Release -t:Install -f net6.0-android

The average of 10 runs on a Pixel 5:

    Before:
    Displayed +1s821ms
    Displayed +1s770ms
    Displayed +1s840ms
    Displayed +1s766ms
    Displayed +1s795ms
    Displayed +1s775ms
    Displayed +1s763ms
    Displayed +1s764ms
    Displayed +1s767ms
    Displayed +1s762ms
    Average(ms): 1782.3
    Std Err(ms): 8.7
    Std Dev(ms): 27.5118156434649

    After:
    Displayed +1s768ms
    Displayed +1s770ms
    Displayed +1s764ms
    Displayed +1s747ms
    Displayed +1s757ms
    Displayed +1s778ms
    Displayed +1s751ms
    Displayed +1s784ms
    Displayed +1s770ms
    Displayed +1s769ms
    Average(ms): 1765.8
    Std Err(ms): 3.62644607172244
    Std Dev(ms): 11.4678293984132

I think this improves startup by ~17ms for this app. Apps with more
views + handlers on startup will have a larger improvement.

### PR Checklist ###

- [ ] Targets the correct branch 
- [ ] Tests are passing (or failures are unrelated)
- [ ] Targets a single property for a single control (or intertwined few properties)
- [ ] Adds the property to the appropriate interface
- [ ] Avoids any changes not essential to the handler property
- [ ] Adds the mapping to the PropertyMapper in the handler
- [ ] Adds the mapping method to the WinUI, Android, iOS, and Standard aspects of the handler
- [ ] Implements the actual property updates (usually in extension methods in the Platform section of Core)
- [ ] Tags ported renderer methods with [PortHandler]
- [ ] Adds an example of the property to the sample project (MainPage)
- [ ] Adds the property to the stub class
- [ ] Implements basic property tests in DeviceTests

#### Does this PR touch anything that might affect accessibility?

No