@jonathanpeppers 2021-11-17
After fixing some typos where I had X & Y flipped, down to one test failure now:
```
IsRunningInitializesCorrectly

Error message
Assert.Equal() Failure Expected: False Actual:   True

Stack trace
   at Microsoft.Maui.DeviceTests.HandlerTestBase`2.<ValidatePropertyInitValue>d__54`1[[Microsoft.Maui.Handlers.ActivityIndicatorHandler, Microsoft.Maui, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null],[Microsoft.Maui.DeviceTests.Stubs.ActivityIndicatorStub, Microsoft.Maui.Core.DeviceTests, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null],[System.Boolean, System.Private.CoreLib, Version=6.0.0.0, Culture=neutral, PublicKeyToken=7cec85d7bea7798e]].MoveNext()
   at Microsoft.Maui.DeviceTests.ActivityIndicatorHandlerTests.IsRunningInitializesCorrectly(Boolean isRunning)
--- End of stack trace from previous location ---
```

Will look into it.