@jonathanpeppers 2021-12-02
@hartez I'm not following how checking `view.Handler == null` would work.

This is set here:

https://github.com/dotnet/maui/blob/963d68563c8499906e64c1d6debc96754a5c1506/src/Core/src/Handlers/Element/ElementHandler.cs#L54-L55

Then a few lines later the properties are updated:

https://github.com/dotnet/maui/blob/963d68563c8499906e64c1d6debc96754a5c1506/src/Core/src/Handlers/Element/ElementHandler.cs#L74

I would like to get rid of `IView.Initialized` too, do you see a way to do it?

We'd probably also like `Initialize()` to be used in cases where maybe a `View` has been removed and re-added. Does that commonly happen while scrolling a `ListView`/`CollectionView`?