@hartez 2021-12-06
> @hartez I'm not following how checking `view.Handler == null` would work.

Yeah, you're right. It wouldn't. I was misunderstanding the order this was all happening in. 

> I would like to get rid of `IView.Initialized` too, do you see a way to do it?
> 
> We'd probably also like `Initialize()` to be used in cases where maybe a `View` has been removed and re-added. Does that commonly happen while scrolling a `ListView`/`CollectionView`?

Looking into this right now.