@hartez 2021-12-06
> > I would like to get rid of `IView.Initialized` too, do you see a way to do it?

Okay, I have a branch that makes this work without the `IView.Initialized` property: https://github.com/dotnet/maui/compare/maui.aar-no-init

The relevant bit is the custom property mapper for Android at the ViewHandler level: https://github.com/jonathanpeppers/maui/compare/maui.aar...dotnet:maui.aar-no-init#diff-a52a8e1cac44a1848e028c1b1730681eb15d2bbdb0405af041d773ffc6c14aca

Here's the setup for it: https://github.com/jonathanpeppers/maui/compare/maui.aar...dotnet:maui.aar-no-init#diff-e802e21561b0e99ffbb055b9ea43d833899fc9f06d91a64c5a9067c9427b7bccR16

It basically just intercepts the key list used by the UpdateProperties method and skips all the properties that can be handled by the mass-update method. And it injects a batch update mapping (with the Initialize method). So this should _only_ be in effect in SetVirtualView then the UpdateProperties method is called, right after a new handler is hooked up. Which should also cover the "removed/reused" situations with virtualization in ListView/CollectionView, since those also call SetVirtualView.

Needs some cleanup and better naming, but it seems to work in my limited testing.