@jonathanpeppers 2021-12-07
Thanks, I'll try your idea tomorrow. If we change that `List<string>` to `HashSet<string>` & use `StringComparer.Ordinal`, should be even better.